+++ lib/services/repository_service.dart
@@ -166,7 +166,7 @@ class RepositoryService {
   /*────────── تنبيه انخفاض المخزون ──────────*/
   Future<void> setAlert({
     required int itemId,
-    required int threshold,
+    required double threshold,
   }) async {
     final db = await _db.database;
